+++ fer_pytorch/logs/finetuning_log.md
