+++ fer_pytorch/logs/resnet34_best/events.out.tfevents.1617275753.elena.12433.0
