+++ tests/test.py
@@ -1,5 +1,4 @@
 import cv2
-
 from fer import FER
 
 fer = FER()
